// types/crm/deals.ts
export interface EnhancedSalesMetrics {
  activeClients: number;
  inactiveClients: number;
  totalCalls: number;
  totalMeetings: number;
  totalEmails: number;
  totalSiteVisits: number;
  followUpNeeded: number;
}